requirements.txt
PyQt5>=5.15
qtawesome>=1.3
opencv-python>=4.8
ultralytics>=8.0
numpy>=1.16.6,<2.4.0

# ── ONNX (required for all export paths) ──────────────────────────────────────────
onnx==1.20.1              # latest 1.x; Ultralytics requires <2.0
onnxslim==0.1.85          # ONNX graph optimiser used by Ultralytics export

# ── GPU inference (CUDA 12.x) ────────────────────────────────────────────
onnxruntime-gpu>=1.20    # ONNX Runtime with CUDA execution provider
tensorrt-cu12>=10.0      # TensorRT 10.x for CUDA 12 (engine export + inference)

# ── CPU inference ─────────────────────────────────────────────────────────
openvino==2025.4.1        # OpenVINO runtime + export
